# api.yml
# This is the OpenAPI specification for your service. It is used to generate the client code.
# To generate the server code, use the Gradle task `openApiGenerate`. The 
# generated code will be placed in `build/generate-resources/main`.
#
# We will evaluate you based on how well you design your API.
# 1. How well it follows REST principles.
# 2. How easy it is to understand and use.
#
# References
# 1. https://swagger.io/specification/
openapi: "3.0.0"
info:
  version: 1.0.0
  title: User Service
  description: |-
    User management service API Contract
  license:
    name: MIT
servers:
  - url: http://localhost:8080
paths:
  /v1/register:
    post:
      summary: Register new user to service
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
                - fullName
                - password
              properties:
                phoneNumber:
                  $ref: "#/components/schemas/PhoneNumberRequest"
                fullName:
                  $ref: "#/components/schemas/FullNameRequest"
                password:
                  type: string
                  minLength: 6
                  maxLength: 64
                  pattern: ^(?=.*[A-Z])(?=.*\d)(?=.*[^\w\s]).{6,64}$
                  example: pAssW0$ds
                  x-oapi-codegen-extra-tags:
                    validate: required,min=6,max=64,_password
                  description: Passwords must be minimum 6 characters and maximum 64 characters, containing at least 1 capital character AND 1 number AND 1 special (non alpha-numeric) character.
      responses:
        '201':
          description: Registration success
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '409':
          description: Phone number already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/login:
    post:
      summary: Log In as registered user, will return JWT 
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
                - password
              properties:
                phoneNumber:
                  $ref: "#/components/schemas/PhoneNumberRequest"
                password:
                  type: string
                  example: pAssW0$ds
                  x-oapi-codegen-extra-tags:
                    validate: required
                  description: Registered user's password.
      responses:
        '200':
          description: Log In success
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/LoginResponse"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/user:
    get:      
      security:
        - bearerAuth: []
      summary: Retrieve user detail
      operationId: getUser
      responses:
        '200':
          description: Retrieve user detail success
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/UserDataResponse"
        '403':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:      
      security:
        - bearerAuth: []
      summary: Update user data      
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              oneOf:
                - required:
                    - phoneNumber
                - required:
                    - fullName
              properties:
                phoneNumber:
                  $ref: "#/components/schemas/PhoneNumberOptRequest"
                fullName:
                  $ref: "#/components/schemas/FullNameOptRequest"
      responses:
        '200':
          description: Update used data success
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/UserDataResponse"
        '403':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '409':
          description: New Phone number already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 
      description: JWT with RS256 algorithm
  schemas:
    ErrorResponse:
      type: object
      required:
        - messages
      properties:
        messages:
          type: array
          items:
            type: string
 
    RegisterResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
        message:
          type: string
 
    LoginResponse:
      type: object
      required:
        - id
        - token
      properties:
        id:
          type: string
        token:
          type: string

    UserDataResponse:
      type: object
      required:
        - phoneNumber
        - fullName
      properties:
        phoneNumber:
          type: string
        fullName:
          type: string

    PhoneNumberRequest:
      type: string
      minLength: 10
      maxLength: 13
      pattern: ^\+62\d+$
      example: '+627000709251'
      x-oapi-codegen-extra-tags:
        validate: required,min=10,max=13,startswith=+62,e164
      description: Phone numbers must be at minimum 10 characters and maximum 13 characters and must start with the Indonesia country code “+62”.
    PhoneNumberOptRequest:
      type: string
      minLength: 10
      maxLength: 13
      pattern: ^\+62\d+$
      example: '+627000709251'
      x-oapi-codegen-extra-tags:
        validate: omitempty,min=10,max=13,startswith=+62,e164
      description: Optional field phone numbers when filled must be at minimum 10 characters and maximum 13 characters and must start with the Indonesia country code “+62”.

    FullNameRequest:
      type: string
      minLength: 3
      maxLength: 60
      x-oapi-codegen-extra-tags:
        validate: required,min=3,max=60
      description: Full name must be at minimum 3 characters and maximum 60 characters.
    FullNameOptRequest:
      type: string
      minLength: 3
      maxLength: 60
      x-oapi-codegen-extra-tags:
        validate: omitempty,min=3,max=60
      description: Optional field full name when filled must be at minimum 3 characters and maximum 60 characters.